cmd: Drop #DUMP by memory address
The ability to dump a Thing by providing a memory location was buggy at
best and at its worst could easily crash the server. As a debugging tool
being able to dump a thing by address was not as useful as initially
thought.

This change removes the ability to dump a Thing by providing an
arbitrary memory address.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>