attr: Switch Door to new Tree debugging scheme
This change switches Door.Dump to use the new Tree output for debugging.
It also standardises the formatting:

  from: "%p %[1]T Exit: %q"
    to: "%p %[1]T - exit: %q"

The formatting of the shared state, shared with the other side of the
door, has also been standardised:

  from: "%p %[1]T Reset: %q Jitter: %q Init: %t Open: %t"
    to: "%p %[1]T - reset: %q, jitter: %q, initially open: %t, open: %t"

The associated shared state, and its associated event.Cancel, are shown
on separate sub-branches. For example:

  |- 0xc000215cc0 *attr.Door - exit: "west"
  |  `- 0xc000163300 *attr.state - reset: "1m0s", jitter: "1m0s", initially
  |     |            open: false, open: true
  |     `- 0xc000187200 event.Cancel

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>