event,attr: Record when Queued events expected to fire
Currently there is no way to know how long a delay is left before an
event will fire. For this reason when creating a copy of an Action,
Cleanup, Reset or Door we cannot copy any associated in-flight events
that are pending.

This change updates event.Queue to also return the expected time the
event will fire. This is not 100% accurate due to system load,
scheduling and processing delays but should be good enough.

As the expected firing time is now known, setting Debug.Events to true
in the configuration file will now include the amount of time an event
was delayed in the delivery message.

For example: "Event delivered (delayed 127.596µs) ...".

The Action, Cleanup, Reset and Door attributes have been updated to
record the event due time and to display the time remaining for an event
or 'Expired' in their Dump information. For example:

  |- 0xc0001049c0 *attr.Reset - after: 10s, jitter: 10s, spawn: true
  |  `- 0xc000063920 event.Cancel - due: 13s
  |- 0xc00004fd40 *attr.Cleanup - after: 10s, jitter: 10s
  |  `- 0x0 event.Cancel - due: expired

Further commits will allow in-flight events to be included in copies of
the attributes.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>