attr,has,frontend,zones: Rename Thing.Copy→Thing.DeepCopy
At the moment the only way to copy a Thing is via Thing.Copy which is
made recursively and includes other Thing such as Inventory content.

This makes it harder than necessary to copy just a specific Thing, which
is going to be needed to fix reset/respawn and unique item duplication
issues.

This change renames Thing.Copy to Thing.DeepCopy and updates its usage
in the attr, has, frontend and zones packages. The individual attribute
Copy functions remain, and still copy a specific attribute. However this
leaves the way clear for a new Thing.Copy implementation that is not
recursive.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>