attr: Clean up Thing.ClearOrigins/Thing.SetOrigins + include disabled
Currently the Thing ClearOrigins and SetOrigins methods clear or set the
origin for a things twice - once when the Inventory is processed and
again when we recurse into the Inventory.

For completeness, when processing the Inventory we should also be
including disabled items. At the moment this has no effect. SetOrigins
is only called when loading zones and disabled items cannot be loaded
into the world currently. The Reset attribute is the only user of
ClearOrigins and Reset is being reimplemented allowing ClearOrigins to
be dropped soon.

This change cleans up ClearOrigins and SetOrigins so the origin is only
set once. It also updates ClearOrigins and SetOrigins to include
disabled items in the Inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>