attr,has: Add Reschedule method to Action, Cleanup & Reset
This change adds the ability to reschedule an Action, Cleanup or Reset
event based on when it is expected to fire. The existing Action, Cleanup
and Reset methods have been refactored into a schedule method which
Action, Cleanup and Reset now call with the after and jitter durations.

A new Reschedule method has been added to Action, Cleanup and Reset
which calls schedule with after set to the period between rescheduling
and the time the event is expected to fire, and zero jitter.

Reschedule has been added to the Action, Cleanup and Reset interfaces.

With hind site the Action, Cleanup and Reset methods should just be
called Schedule - maybe setup a has/event interface for all event
related attributes? Work for another commit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>