attr: Fix respawn, in-flight events and duplication bugs in Reset
At the moment there are three bugs related to item spawning:

  1. Spawnable items in a spawnable container do not have their origin
     set/cleared properly and if the (now incorrect) origin is disposed
     of it causes a nil panic. This is dependant on whether the
     container was in its original spawnable state or was picked up and
     is now non-spawnable.

  2. If a spawnable item is taken from a spawnable container (and the
     container is in its original spawnable state) and the container is
     then taken and the original is taken out of play, then the original
     spawnable item that was waiting for a reset never resets as the
     in-flight event is lost and the item stays disabled.

  3. If a unique item is put into a spawnable container (and the
     container is in its original spawnable state), then taking the
     container causes the unique item to be duplicated - which should
     not happen as unique items should remain unique.

This change reimplements the Reset.Spawn method and adds a new
Reset.spawnInventory method to handle resetting and respawning of items.
The new spawnInventory is a custom implementation of Thing.DeepCopy that
caters for moving unique and non-spawnable items to spawned containers
and spawns copies of spawnable items. For disabled items waiting for a
Reset event the spawnInventory method will reschedule the events. For
newly spawned items the spawnInventory method will also update the
origins. These changes therefore fix the three issues identified.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>