attr/has: Add Action,Cleanup,Reset Suspend method
At the moment Action, Cleanup and Reset events cannot be suspended and
resumed. This is because only the due timestamp is recorded. To resume
an event we also need to know the time period remaining before an event
should fire.

Currently a Reset event can be saved and restored - but it must be
stopped at the point of saving and start when reloaded. This causes an
issue if a player enters the game and then exits to the menu - events
will continue to run. The event cannot be suspended and resumed as we
don't know how much longer we have to wait before the event fires.

This change renames the due time to dueAt and adds a dueIn period. It
also adds a new Suspend method for suspending an event. The Resume
method has been updated to restart an event using the new dueIn.

When an event is started it sets the Cancel channel and dueAt. The Abort
method will close the channel and clear the dueAt value. If Suspend is
called the channel is closed, dueAt cleared and the period remaining is
recorded in dueIn. If Resume is called a new event is started with the
delay set to the dueIn value.

The Marshal methods have been updated to record the dueIn value as
DUE_IN, as when saved the event is suspended. The Unmarshal methods have
been updated to accept either DUE_IN or DUE-IN.

The Dump methods have been updated to show the source of the due value
as either 'at' or 'in'.

The Copy methods will produce a copy with the event suspended and dueIn
value recorded if an event is in-flight or suspended when copied.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>