attr/cmd: Update meaning of what is collectable
Due to recent changes and fixes it is now possible for a player to
obtain a spawnable item in their inventory, usually by obtaining a
container that has a spawnable item. The definition of what is
collectable needs to be updated to cater for this.

This change updates Thing.Collectable to be anything that is unique and
is not a player. Due to this change the junk.dispose method can no
longer use Thing.Collectable to determine if specific items can be freed
for garbage collection. Instead specific checks must be made.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>