cmd: Cleanup Reset event when taking shortcut in $RESET
Currently in the $RESET command if the reset is not seen a shortcut is
taken that does not cleanup the event that triggered the reset. This can
lead to the event keeping a reference to the Cancel channel - which will
eventually be released and garbage collected.

This change explicitly cleans up the event by calling Reset.Abort in the
shortcut path, just as the main path does.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>