cmd: Save player's disabled items in SAVE command
Currently when saving a player and processing their inventory disabled
items - those with in-flight Reset events - are not saved. Now that
players can keep containers with spawnable items in them this is an
issue.

This change updates the save.inventory method to also save disabled
items when saving a player.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>