attr: Record disabled items in Inventory.Marshal
Currently references for disabled items are not recorded when an
Inventory is Marshaled. This is a problem now that players can keep
spawnable items that may be waiting for a Reset event to occur.

This change also records disabled items. However it is not currently
possible to identify disabled items by looking at the Inventory field in
a *.wrj file - you have to check for a Reset field that has a due-in
property. This is not ideal and will be cleaned up in a later commit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>