frontend,quit: Resume/Suspend Reset events entering/leaving game
Due to recent changes it is now possible for player's to have spawnable
items in their inventory. These items may have an in-flight Reset
events. These events should be resumed only when the player enters the
game and suspended when the player exits the game (or when initially
loaded). At the moment players can enter the game, initiate a reset, and
exit to the main menu where they can wait for resets to occur before
re-entering the game.

This change adds a game.resumeResets method to resume any suspended
Reset events when entering the game. It also adds a quit.suspendResets
to suspend any in-flight Reset events when leaving the game.

Adding these methods to frontend.game and cmd.quit is probably not the
best place for them, but will do for now. Should they be moved to
attr.Inventory or attr.Reset?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>