attr,cmd,frontend,zones,docs: Improve disabled item loading
This commit is a follow on from 9255173: "attr: Record disabled items in
Inventory.Marshal". Currently the only way to identify if an item should
be disabled when loaded from a *.wrj file is to look for a suspended
Reset event - a RESET with a DUE_IN/DUE-IN period set. This causes all
items to be loaded enabled, and then disabled if a suspended Reset is
found.

This commit allows disabled items to be specified on INVENTORY and
LOCATION fields by prefixing them with an exclamation mark '!'. Such
items are loaded disabled, and will stay disabled if there is no
suspended Reset to bring them into the game. It is planned that other
means of bringing the item into play will be added in the future.

The Inventory.Unmarshal method has been updated to write out disabled
references with a leading exclamation mark '!'.

The save.fixInventory method for the SAVE command has been updated to
understand disabled references.

The zone.linkupInventory and zone.linkupLocation in the zone loader have
been updated to understand disabled references.

The frontend.log.assemblePlayer method has been updated to understand
disabled reference when loading a player.

The frontend.game.resumeResets has been simplified as it no longer has
to scan enabled items for a Reset and then disable them. Setting the
origin for a player and their inventory only needs to be performed once,
when the player is loaded. Due to this the setting of origins has been
moved to the frontend.login.assemblePlayer method.

The documentation in docs.zone-files.txt has been updated with details
on disabled references and use of DUE_IN/DUE-IN for Reset fields.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>