attr,has,cmd: Fix disposal of copied spawnable (added Reset.spawned)
Since the recent rewrite and fixes to item spawning it is possible to
get a collectable, spawnable item inside a container. However it now
becomes impossible to tell the difference between the original spawnable
items and the copied ones. This causes panics at random points
throughout the code as junk.dispose cannot correctly free items
resulting in stale inventory references and improbable Reset attempts.

This change adds a new spawned field to Reset which is set by the
Reset.spawnInventory method by calling Reset.Spawned. This flag can be
checked by calling the new Reset.IsSpawned method which returns true if
the Thing has been spawned and is a copy, else false. This allows us to
determine the difference between original spawnable and copied spawnable
items.

The junk.dispose method has been updated to use the new Reset.IsSpawned
method.

Note that the Reset.Spawned has been added to the has.Reset interface as
we can't access Reset struct fields through an interface. It shouldn't
really be exposed and will be cleaned up when the attributes are sorted
out into their own packages.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>