cmd: Fix data race in state locking
Currently the newState function will get the actors location, store it
in state.where and then add the outermost Inventory to the locks list.
The locks are then acquired in the state.sync method. It is possible
that the Inventory hierarchy can change, without state.where changing,
between state.newState being called and state.sync being called
resulting in the wrong lock being acquired and resulting in a data race.

This situation is especially true where the actor is in a container that
moves. The actor stays in the container, so state.where does not change,
but the container moves to different Inventory - resulting in the
Inventory hierarchy changing. With the Inventory hierarchy changing the
locks acquired are no longer invalid and the required locks are not
acquired - resulting in a data race. An example of this situation is a
reset happening in a container that is carried by a player - the Reset
event fires and the player moves between state.newState and state.sync
being called for the event handling.

This change moves the setup of state.where into the state.sync method.
In state.sync, once the lock for state.where has been taken we
revalidate that the lock is still valid for the actor. If the Inventory
hierarchy has changed, invalidating the lock, all of the locks are
released, state.where is updated and locking is tried again - until the
locks can be acquired and validated successfully.

Note that it may require a number of attempts to successfully acquire
and validate the required locks. This should not be severe enough to
warrant a more complex solution - testing on a very busy server results
in a worse case of less than ten retries which includes normal command
relocking. If this method does become a problem we could add a RWMutex
to the state package. Normally we acquire the lock for reading before
calling sync, if we have to retry too many times we acquire the lock for
writing which will 'stop the world' and pause all other command
processing while we take our locks, then we release the write lock.

An alternative approach would be to add the actor's new location to the
locks we want and lock all of the locations. However if the actor's
position keeps changing we could end up acquiring a lot of locks
unnecessarily which would impact performance, especially if we have to
release and reacquire the locks due to command processing as well, such
as SNEEZE or simply moving using N, E, S, W, etc.

A number of comments for state have been updated for the new processing
and clarification.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>