cmd: lock origins of disabled items in junk.lockOrigins
With the recent spawn fixes it is now possible that disabled items
need to be reset or disposed of. Currently the origins of disabled items
are not locked resulting in data races.

This change updates junk.lockOrigins to also lock the origins of
disabled items.

Note that this should only effect unique, spawnable items as non-unique
items never leave their origin.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>