attr,has: Store 'Ref' field in Thing when unmarshaling
Currently the Ref field from a record jar is ignored when unmarshaling a
Thing. However, there are situations were this value would be useful.

This change adds a new ref field to the Thing struct for storing the
unmarshaled Ref value. The value is not unique and can be the empty
string. The Dump method has been updated to show the current value.

The Thing.Unmarshal method has been updated to no longer ignore the Ref
field, but to store it instead.

The Thing.Copy and Thing.DeepCopy methods have been updated to also
include the Thing.ref field value in the copy.

A new Thing.Ref method has been added to access the Thing.Ref field
through the has.Thing interface - which has been updated to add the Ref
method.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>