attr,has: Add Inventory.SearchByRef method
This change adds a new method SearchByRef to Inventory to find a Thing
by its reference - by calling Thing.Ref on each item on the Inventory.
The search is not recursive into other Inventory.

Ideally the current Inventory.Search method should be refactored to
Inventory.SearchByAlias, but than can wait until the attributes are
reworked.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>