cmd: Update EXAM/EXAMINE to use bodies instead of players
Currently the EXAM/EXAMINE command only describes what is being held,
worn or wielded if the participant is a player. However, this should
work for anything with a body such as mobiles and items - such as a doll
wearing a tiny dress.

This change updates the testing for a Player to testing for a Body
instead.

Updated pkg_test.go so that test players are created with a body.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>