attr,has: Add post-unmarshaling Load & pre-marshal Save hooks to Thing
This change adds Load and Save methods to Thing so that attributes have
a hook to perform additional processing when a Thing is unmarshaled or
marshaled.

The Load method provides a post-unmarshaling hook, the Save method provides
a pre-marshaling hook.

For attributes that need to perform additional processing and take
advantage of the hooks they should implement a non-exported load and/or
save method. For example:

  func (n *Name) load() {
    // Additional processing...
  }

  func (n *Name) save() {
    // Additional processing...
  }

When the Thing.Load method is called any load methods implemented by
attributes will be called, while Save will call any save methods
implemented by attributes.

Load will be called after the Thing is unmarshaled, a deep copy is made
and origins are set - this also applies to inventories of the Thing
recursively, depth first. This allows an attribute to call its Parent
method and access other attributes of the Thing - which cannot be done
during unmarshaling.

Save will be called before the Thing is removed from the game world and
marshaled. Save will process inventories of the Thing recursively, depth
first.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>