has,attr: Add new Holding,Wearing and Wielding attributes
This change adds new Holding, Wearing and Wielding attribute that
represents the items being held, worn and/or wielded by a Thing. The
current implementation is enough so that players and zones can be loaded
and items will be held, worn and/or wielded according to the Holding,
Wearing and/or Wielding fields in a record jar.

Documentation added to docs/zone-files.txt for new record jar fields.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>