data: Update city guard to use items (helmet,cuirass,boots,shortsword)
This change adds a Body to the city guards. It also gives the city
guards a helmet, cuirass and boots to wear and a shortsword to wield.

NOTE: While the items will initially be used when the zones are loaded
resetting a guard who has had items removed will not use the items
again. This issue is currently being worked on.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>