data: Update flower girl to wear dress + hold bunch of flowers
This change adds a Body to the flower girl. It also gives the flower
girl a dress to wear and a bunch of flowers to hold.

NOTE: While the items will initially be used when the zones are loaded
resetting the flower girl when items have been removed will not use the
items again. This issue is currently being worked on.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>