cmd: Handle new WAIT value for Reset attribute in $RESET
This change updates the $RESET command to delay resetting of containers
where the Reset attribute has WAIT specified.

Containers only delay resetting while they have disabled items in them
waiting for a reset.

For the current implementation of WAIT if a container has disabled items
and needs to wait it will reschedule its reset event. It will keep doing
this until all items in the container are ready to reset.

Due to the way unique items are implemented a container will only wait
for unique items if they already have a reset event pending. This is
because unique items are moved around and only disabled when waiting for
a reset. Spawnable items are copied and the copies put in play while the
originals are disabled and a reset scheduled.

Therefore a container can only wait on a disabled unique item when it
happens to be waiting for a reset. Original spawnable items will always
be in their container - in either an enabled or disabled state and so
can always be waited on.

The documentation in docs/zone-files.txt has been updated for Reset and
has more detail on WAIT.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>