frontend,cmd: Fix data race when checking if player left world
The commit cd5694a5: "frontend: Use Locate.Where as exiting signal in
frontend.game" introduced a subtle data race in the game.process method.
Specifically in the code:

  l := attr.FindLocate(g.player)

  // Only pass command to game parser if still in the world
  if l.Where() != nil {
    cmd.Parse(g.player, string(g.input))
  }

  // If no longer in the world switch to frontend main menu
  if l.Where() == nil {

The data race is due to the fact we test for "l.Where() != nil" and then
call cmd.Parse which does all of the locking via sync. However, the
check and call to cmd.Parse are unrelated. It is therefore possible for
the following to occur:

  if l.Where() != nil {
    // <--- Player gets killed here and l.Where() now returns nil
    cmd.Parse(g.player, string(g.input))
  }

This typically manifests as a nil pointer panic in the MOVE command when
testing.

This change modifies the cmd.state.sync method to check if the actor's
location is nil, if is is nil sync gracefully exits.

The second check for "l.Where() == nil" after calling cmd.Parse is safe.
Once a player's location is set to nil it can only be set to a valid
location by re-entering the game, which involves exiting the game
frontend and going through the menu frontends.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>