cmd: Fix sync s.where nil check for locations (fixes Events)
The commit 1793cf5: "frontend,cmd: Fix data race when checking if player
left world" was incorrect and prevents events such as clean-ups and
resets from happening.

This is because locations are at the top of the Inventory tree and
do not have a Locate attribute and so Locate.Where() will always return
nil. Commit 1793cf5 checks for s.where being nil, and in the case of
locations stops processing when it shouldn't.

This change fixes commit 1793cf5 by also checking for an Exits
attribute, which is what makes a location a location.

Note that the fix could check for not have a Locate attribute or
specifically checking for a Player attribute. However, checking for not
having an Exit attribute seems safest.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>