attr,has: Add WAIT to Reset attribute
This change adds processing to the Reset attribute for a new WAIT
key/value pair on the Reset field in recordjar files. The value is
expected to be true or false, defaulting to false if not specified. If
WAIT is specified by itself it is equivalent to WAIT→true.

Example usage:

  Reset: AFTER→1m JITTER→1m SPAWN WAIT

A new Reset.Wait method has been added to query the value of WAIT and is
exposed in the interface.

By itself this change only marshals, unmarshals and allows the querying
of the WAIT value, but does nothing else with it. Functionality needs to
be added to the $RESET command to have any effect.

It should be noted that NewReset takes an additional wait parameter. The
get and take tests have been updated to pass false in the wait parameter
to NewReset which preserves backwards compatible  behaviour,

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>