attr: Add Body.RemoveAll and use in Body.Free
This change adds a RemoveAll method to the Body type and interface. The
method will unconditionally remove all Thing currently in use and clear
the associated slot usage flags.

The Body.Free method has been updated to use RemoveAll in place of a
simple loop in case RemoveAll needs to do additional bookkeeping in the
future.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>