attr: Add reset hooks to holding/wearing/wielding
This change adds resetHook methods to the Holding, Wearing and Wielding
attributes so that items can be used by a Thing when it resets.

NOTE: Unlike the load hooks the reset hooks are silent. This allows a
container to reset and only use items that are available - which may not
be all of the items if some have not reset.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>