attr,has,cmd,frontend,zones: Bring load+save hook inline with reset
The naming of ResetHooks and resetHook seems more sensible than the
current Load/Save and load/save hook methods. This change renames the
Load/Save methods LoadHooks/SaveHooks, and renames the load/save methods
loadHook/saveHook.

This change also updates the comments for the changed methods to be more
accurate and concise.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>