zones: Add zone file record jar pre-processor for @refs
This commit adds a simple preprocessor for @refs in zone files. An @ref
can be used to reference the definition of a field at another reference,
hence the feature is called '@ref' or 'at ref'.

For example: "Body: @STD_BODY" - the definition for Body is at the Body
field in the STD_BODY record (Ref: STD_BODY).

A more complete example:

  %%
       // Standard body definition
     Ref: STD_BODY
    Body: ANKLE→2 BACK→1 CHEST→1 EAR→2 ELBOW→2 EYE→2 FACE→1 FINGER→8
          FOOT→2 HAND→2 HEAD→1 KNEE→2 LOWER_ARM→2 LOWER_LEG→2
          LOWER_LIP→1 MOUTH→1 NECK→1 NOSE→1 PELVIS→1 SHOULDER→2 THUMB→2
          UPPER_ARM→2 UPPER_LEG→2 UPPER_LIP→1 WAIST→1 WRIST→2
  %%
       // Short durations
     Ref: SHORT
   Reset: AFTER→1m JITTER→1m
  Action: AFTER→30s JITTER→30s
  %%
       // Long durations
     Ref: LONG
   Reset: AFTER→10m JITTER→10m
  Action: AFTER→2m JITTER→2m
  %%
          // Standard inventory
        Ref: STD_INV
  Inventory: MUG ID_CARD
  %%
          // Supplemental inventory
        Ref: BOSS_INV
  Inventory: CLIPBOARD PEN
  %%
          // Standard actions
        Ref: ACTIONS
   OnAction: $ACT hums to themselves.
           : $ACT takes a sip from their mug.
           : $ACT fiddles with their ID card.
  %%
          // Supplemental actions
        Ref: BOSS_ACTIONS
   OnAction: @ACTIONS
           : $ACT taps their clipboard with a pen.
           : SAY I can't see you working.
  %%
        Ref: M1
       Name: Alice
      Alias: ALICE
       Body: @STD_BODY
      Reset: @LONG WAIT
     Action: @SHORT
   OnAction: @BOSS_ACTIONS
  Inventory: @STD_INV @BOSS_INV

  This is Alice.
  %%
        Ref: M2
       Name: Bob
      Alias: BOB
       Body: @STD_BODY
      Reset: @SHORT
     Action: @SHORT
   OnAction: @ACTIONS
  Inventory: @STD_INV

  This is Bob.
  %%

The example defines a standard Body, some standard LONG and SHORT Reset
and Action fields and some standard inventories and actions.

The standard body is used 'Body: @STD_BODY' for both Alice and Bob and
specifies that the definition for Body can be found at the record with a
reference of STD_BODY.

Only the @ref is replaced with the data from the referenced field. This
means other data can appear either side of the @ref. For example Alice
has a Reset field that has an @ref of @LONG followed by WAIT. This
will expand to:

  Reset: AFTER→10m JITTER→10m WAIT

Multiple @refs may be specified on a field. For example Alice's
inventory is specified as '@STD_INV @BOSS_INV' which will expand to:

  Inventory: MUG ID_CARD CLIPBOARD PEN

The definition pointed to by an @ref may itself contain one or more
@refs. For example the definition for Alice's OnAction field is
@BOSS_ACTIONS and the definition for BOSS_ACTIONS includes a reference
to @ACTIONS. When expanded Alices's actions are therefore:

 OnAction: $ACT hums to themselves.
         : $ACT takes a sip from their mug.
         : $ACT fiddles with their ID card.
         : $ACT taps their clipboard with a pen.
         : SAY I can't see you working.

The definitions of Refs and @Refs can appear in any order in the Jar,
that is @refs do not have to be defined before they are used in the Jar.

After all expansion of @refs has been done Alice and Bob would be
defined as:

  %%
        Ref: M1
       Name: Alice
      Alias: ALICE
       Body: ANKLE→2 BACK→1 CHEST→1 EAR→2 ELBOW→2 EYE→2 FACE→1 FINGER→8
             FOOT→2 HAND→2 HEAD→1 KNEE→2 LOWER_ARM→2 LOWER_LEG→2
             LOWER_LIP→1 MOUTH→1 NECK→1 NOSE→1 PELVIS→1 SHOULDER→2
             THUMB→2 UPPER_ARM→2 UPPER_LEG→2 UPPER_LIP→1 WAIST→1 WRIST→2
      Reset: AFTER→10m JITTER→10m WAIT
     Action: AFTER→30s JITTER→30s
   OnAction: $ACT hums to themselves.
           : $ACT takes a sip from their mug.
           : $ACT fiddles with their ID card.
           : $ACT taps their clipboard with a pen.
           : SAY I can't see you working.
  Inventory: MUG ID_CARD CLIPBOARD PEN

  This is Alice.
  %%
        Ref: M2
       Name: Bob
      Alias: BOB
       Body: ANKLE→2 BACK→1 CHEST→1 EAR→2 ELBOW→2 EYE→2 FACE→1 FINGER→8
             FOOT→2 HAND→2 HEAD→1 KNEE→2 LOWER_ARM→2 LOWER_LEG→2
             LOWER_LIP→1 MOUTH→1 NECK→1 NOSE→1 PELVIS→1 SHOULDER→2
             THUMB→2 UPPER_ARM→2 UPPER_LEG→2 UPPER_LIP→1 WAIST→1 WRIST→2
      Reset: AFTER→1m JITTER→1m
     Action: AFTER→30s JITTER→30s
   OnAction: $ACT hums to themselves.
           : $ACT takes a sip from their mug.
           : $ACT fiddles with their ID card.
  Inventory: MUG ID_CARD

  This is Bob.
  %%

NOTE: In the examples the Inventory items (MUG, ID_CARD, CLIPBOARD and
PEN) have not been defined for brevity.

The docs/zone-files.txt document has a new scetion explaining @refs with
more examples.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>