recordjar,docs: Fix named free text field & free text section usage
At the moment it is possible, but undocumented, to have a named
Description field, as opposed to an unnamed free text section that is
automatically named Description. For example the following two
definitions are equivalent:

  %%
  Description: This is some text!
  %%
  This is some text!
  %%

However, if a named field is used the resulting field in the jar will
erroneously contain the field name in the text:

  record["DESCRIPTION"] = "Description: This is some text!"

If a named Description field is used with an unnamed free text section:

  %%
  Description: This is

  some text!
  %%

Then the content are combine, but with the field name and extra newline
separators erroneously included:

  record["DESCRIPTION"] = "Description: This is\n\nsome text!"

If the named Description field is continued the leading white-space is
erroneously included along with the newline:

  %%
  Description: This
        is

  some text!
  %%

Results in:

  record["DESCRIPTION"] = "Description: This\n      is\n\nsome text!"

In all of these cases the result should be:

  record["DESCRIPTION"] = "This is some text!"

This change causes named Description fields to be treated the same as
any other field. The free text section is now collected in an internal
field with the name "FREE TEXT". This cannot accidentally be used by a
user as field names cannot contain white-space in them. Before the
record is placed in the jar the named Description field and internal
free text section are examined:

  - If there is a named Description field and an unnamed free text
    section then the section is appended to the named Description field.
  - If there is only an unnamed free text section it becomes the data
    for the named Description field.
  - If there is only a named Description field it is left as is.

This corrects all of the above issues, which are covered by new tests.
The documentation in docs/zone-files.txt has been updated to include the
Description field.

NOTE: For record jars the name of the free text block can be specified
per jar. For WolfMUD zone files this is 'Description'. This change
applies regardless of the field name chosen.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>