recordjar: Refold free text section when writing
At the moment free text sections are folded when being written. However
it is possible for a free text section to be read and not unfolded. If
this free text section is then written it could have some weird
formatting issues.

This change will unfold the free text section before folding it again to
make sure the formatting is correct.

Test added to make sure free text sections are refolded when written.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>