recordjar: Add optional ordering to Write
Currently when a record jar is written the fields for a record are in
alphabetical order.

This change allows a slice of field names to be passed to Write to
specify a different ordering. If nil or an empty slice is specified the
ordering will be alphabetical as it is now. If a slice containing names
is specified then the fields within a record will be ordered according
to the passed slice. For fields not listed in the passed slice, they are
ordered alphabetically and added to the end of the record - after the
specifically sorted fields.

Tests and the SAVE command have been updated to pass a nil list.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>