Added Makefile (Linux/Bash only) and bin directory
This is the Makefile used for development and testing. It is intended
for Linux and uses Bash, no Windows equivalent is available yet :(

Tested on Debian (testing/bullseye).

The Makefile is not required, 'go build -o bin/ ./...' works as well :)

The bin directory has a .gitignore to ignore everything in it.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>