Add build environment for making WolfMUD releases
This is the build environment used to create WolfMUD releases. It
contains a few helper scripts and a Makefile. It is intended for Linux
and uses Bash, no Windows equivalent is available yet :(

Tested on Debian (testing/bullseye).

The only thing not included (yet!), and not required, is 'jt' the "Just
Text" static site generator. All it does is take the index.txt and
WolfMUD-*.txt files and turns them into the download pages as seen on
the WolfMUD.org website. Without 'jt' an unlinked full-index.txt is
created instead using a little sed magic :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>