build: New addAcct for creating new player accounts
This commit adds a new addAcct Bash script that creates new accounts and
writes the player files into ../data/players/ or a directory specified
with the -d flag. This is intended for bulk account creation when
testing using a large number of bots/robots, but may also be used to
create an ordinary player account - e.g. if a player is having issues
registering for some reason.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>