tools: botrunner discard should only exit on errors
There is a bug where the botrunner discard exits on any error, including
timeouts. This can cause discard to terminate early, leaving the read
buffers to fill and the network connection for a bot to eventually fail.

This commit adds a longer timeout deadline to the discard read and
prevents discard from exiting on timeouts.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>