tools: Pause botrunner logins for 10 seconds every 2048 bots
When a large number of bots log into the server all at once it can cause
a high level of lock contention in the starting locations. This commit
adds a ten second delay every 2048 bot logins to give the crowds of bots
a chance to dissipate and leave the starting locations.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>