comms: Make network writes to clients asynchronous
There is currently an issue with slow clients or a congested network
that can cause delays in state.messenger when processing player
commands. This in turn delays further commands and so on, snowballing
until the server grinds to a halt.

This commit makes network writes to the client asynchronous and prevents
delays in processing player commands due to networking. The comms.Write
method has been renamed comms.write and writes directly to the network
connection. A new comms.Write method has been added that puts message
into a backlog channel. A goroutine runs the comms.backlogWriter which
takes messages from the backlog and passes them to comms.write to send
them to the client.

The backlog is currently sized to hold up to 48 messages - two screens
worth of messages on an 80x24 screen. If the backlog becomes full the
oldest messages are dropped to make room for new messages. This is
preferable to dropping clients. In most cases messages will only be
dropped if the client's connection is very slow, the server is
generating a lot of messages all at once or if the server is overloaded
and cannot handle all of the network traffic.

The 'cost' of this change is an additional goroutine per player, the
storage taken up by the backlog channel and additional allocations due
to a copy of the message having to be written to the backlog.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>